Fix test alternate_thread_diversion with older lldb versions.
In old versions of lldb the output format of the `print` command was
different, it shows a value history identifier.

Therefore the output of `print get_value()` is not recognized
in the alternate_thread_diversion test and the test times out.

lldb version 14.0.6
'print' is an abbreviation for 'expression --'
(rr) print get_value()
(int) $0 = 1

lldb version 18.1.3
'print' is an abbreviation for 'dwim-print --'
(rr) print get_value()
(int) 1