Bring up all runners in parallel with a single Lambda call
The Lambda call waits for all runners to be registered with Github before it returns.
It also downloads the latest Github Actions runner version automatically and blocks automated updates
so the runner doesn't go offline temporarly immediately after starting up. Hopefully this will avoid
situations where the runner is running but the job hasn't been picked up.

If we have reliability issues with machines not starting up properly or not accepting
jobs this should make it easier to debug and fix in future, since the Lambda
can log the state of the system.

I've also changed the runner to mount a 32G tmpfs on /tmp, and run the runner
and its working data in /tmp, which should help with some space issues.

And I've moved "update the distro packages" out of the cloud-init script into
`github-actions-build.sh`. It's hard to debug stuff in the cloud-init script
so we want to minimize it.