Only check for PROT_{READ,WRITE,EXEC} in assert_segments_match()
On arm64, mapping protection flags can include PROT_BTI and PROT_MTE,
but they are not shown in /proc/pid/maps, so if a mapping has one of
these the assert_segments_match() check will fail. Fix it by masking
out flags that are not shown in /proc/pid/maps when doing the check.