Defer unmapping syscallbufs of dead procs
If two processes share an address space (via vfork(2) or clone(2)
CLONE_VM), and one processes calls execve(2), the process's syscallbufs
will remain mapped in the shared address space. At the moment, we
attempt to unmap this after execve(2) by finding a stopped process in
the shared address space, and using it to do the unmapping. If there is
no such process, though, the buffers are leaked.

This patch changes that situation to ensure that such syscallbufs are
unmapped in all circumstances. This is done by recording that such
unmapping must happen after execve(2), but deferring the unmapping until
a suitable process is scheduled and in a state to perform the unmapping.