Don't assert on ticks overshoot during record on AMD
On AMD, we regularly overshoot the programmed ticks counter.
We print this as a warning earlier in the function. However,
the warning is ineffective, because we have an assert with the
same condition later in the function. Adjust the check for
that error condition to only fire during replay, not record
(where it is supposed to be non-fatal).