Add test case for #3831
This one depends on the timing of a SIGCHLD (needs to happen right when
resuming the parent task, so that it gets delivered in the AutoRemoteSyscall
for the syscallbuf cleanup), so it's somewhat non-deterministic. However,
but exiting fast enough (and the dynamic linker is not), it's possible
to trigger this condition quite reliably, so add an executable that does
that.