Convince compiler to use constructor MemoryRange(addr, end).
A default build (no additional cmake options) in a 32-bit Debian Bookworm
was observed to use the constructor MemoryRange(addr, size)
instead of MemoryRange(addr, end).
This triggered a crash of the chaos_oom test, because it tried to
use the last part of the address space, which results later in
following fatal message, because end_ wraps over to the value 0.

[FATAL src/MemoryRange.h:20:MemoryRange()] start_ <= end_