Accept `restart_syscall` for the result of `sigreturn`
I have seen the following assertion failure:

-> Assertion `syscallno == ret` failed to hold. Interrupted call was nanosleep and sigreturn claims to be restarting restart_syscall

I don't understand how this could happen --- nanosleep should never be restarted after sigreturn, it should return `EINTR` instead. If you interrupt `nanosleep` with a signal with default disposition (no handler), the kernel restarts it with `restart_syscall`, so I can imagine that if the kernel restarted that after a signal handler, we would get this assertion. Still, it shouldn't happen. I tried various tests and couldn't reproduce it.

This kernel may have been patched, so for now my best theory is that a patch or a bug allowed the restart to happen. We should simply relax the assertion to allow `restart_syscall` here.