pack: Restore behavior with no passed trace dirs
This is a bit of an embarassing bug-fix, but as pointed out in
https://github.com/rr-debugger/rr/pull/3834#issuecomment-2380622819,
I accidentally broke `rr pack` without trace dir specified (which is
supposed to pack the latest dir). We do actually have a test for it,
which does fail for me locally (which I didn't notice - sorry),
but not on CI. I'm not entirely sure why it doesn't fail on CI, but
I imagine we might be copying more things in the CI environment, rather
than cloning/linking them. To try to catch similar regressions in the future,
this also adds an extra check to the test to make sure that there is
some `mmap_pack_*` file in the trace after packing. I think we should
always be packing at least libc or ld-linux, so hopefully there
shouldn't be any false positives there.