Try to fix intermittent failure in alternate_thread_diversion test
We sometimes fail like this:

[FATAL src/GdbServer.cc:1379:handle_exited_state()] Received continue/interrupt request after end-of-trace: 6

with LLDB only. The test script has finished the test and we exit the test while racing with the test program itself processing SIGKILL and exiting. (We don't stop for SIGKILL in this test.) It looks like LLDB for some reason sends an interrupt packet and we receive it after tracee exit.