Don't allow chaos mode to map memory such that it merges with the stack
When running a program under rr, we remove MAP_GROWSDOWN from the stack
mapping. If rr in chaos mode subsequently places a mapping immediately
above the stack, it will actually get merged into the same VMA.

This is a problem because glibc actually looks at /proc/self/maps to dig
out the stack top in pthread_attr_getstack; this will be wrong if the
stack VMA is actually merged with another one.

Co-authored-by: Rian McGuire <rian@rian.id.au>