Add CPUID lists links to PerfCounters.cc
Due to the apparent and constant churn when adding "confirmed working" CPUIDs, I 
decided to clean up the comments here. Take a hint, this means many CPUs aren't 
explicitly added to the list of known processors.

Updated Intel's link and gave directions where exactly to find CPUIDs.

AMD doesn't appear to have a single source of truth for those, as they are 
scattered in individual revision guides and are hard to find.