Properly track deferred syscall patching with the syscall event
This fixes #3880 by moving the deferred syscall patching flag from
Task to the corresponding syscall event. Both to avoid the assertion
in that issue and for performance (since unpatched syscalls are traced),
it is important that we (attempt to) patch the correct syscall ip.

Additionally, we add the ability for tests to make sure that a particular
syscall was actually patched. In particular, our logic for bailing out
when another task is in the patch region was accidentally preventing
all future attempts at patching the same region, even if doing so
would have succeeded. With that fixed up, the new test will make sure
this doesn't regress.