Detect when 5 level paging was used during recording but is not available during replay.
Support (or lack thereof) for 56 bit addresses is a trace portability hazard
that we want to detect up front. We record whether or not 5 level paging was
used in the trace, which is trivially observable from the addresses that
pass through AddressSpace::map. Then during replay of a trace that uses 5
level paging we attempt to map something above the standard 47 bit address
space, and if it fails, we die with a useful error message.