Patch `__aarch64_have_lse_atomics` after `execve` if possible
https://github.com/bminor/glibc/commit/018f0fc3b818d4d1460a4e2384c24802504b1d20
added atomic operations early in `ld-linux`. These are compiled with outline-atomics
but the `__aarch64_have_lse_atomics` is not set because the constructor has
not run (maybe it never does in `ld-linux`). So let's set it ourselves.

`__aarch64_have_lse_atomics` is not an exported symbol, so we have to find it
ourselves by disassembling `__aarch64_ldadd4_relax`. That's a very simple function
so this should be robust enough.