Fix PTRACE_INTERRUPT to actually work if the tracee is not already stopped. (#3923)
The current code is completely untested because it will always call
syscall_state.emulate_result twice and fatally assert. The ptrace_seize test
does not yield between PTRACE_CONT and PTRACE_INTERRUPT, so rr will not
switch to the emulated ptracee and resume execution, and the PTRACE_INTERRUPT
will always find the emulated ptracee already stopped. Additionally, the test
does not wait on the emulated ptracee again after interrupting it, so the
following assertion about the value of status is bogus.

Fix all of that by adding the appropriate waitpid, tweaking the status
assertion, duplicating the PTRACE_CONT/PTRACE_INTERRUPT sequence but with
an intervening sched_yield() to force the emulated ptracee to resume, and
fixing rr's PTRACE_INTERRUPT emulation to emulate_result only once and to
place an emulated stop on the emulated ptracee.