Mimic kernel signal coalescing behavior for synthetic SIGCHLDs. (#3927)
On the real kernel multiple SIGCHLDs (and all non-rt signals, of course) are
coalesced and only a single SIGCHLD is delivered. It's important for us to
copy this behavior here because RecordTask::has_other_actionable signals does
not know about the signals that effectively exist in
emulated_[ptrace_]SIGCHLD_pending. If we allow multiple emulated SIGCHLDs and
those SIGCHLDs happen to interrupt e.g. a recv(2) the EV_SIGNAL_DELIVERY case
in RecordSession::signal_state_changed will set up the syscall restart after
the first signal only to immediately take a second one, and when we go to replay
the second signal we will find the registers have diverged.