Copy `last_task_tuid` when cloning a `ReplaySession`
The bug is obvious but it's hard to test.
When `last_task_tuid` is not initialized properly in a cloned session,
`ReplaySession::setup_replay_one_trace_frame()` can make a spurious call to
`Task::will_schedule()` which can update rseq state at a point where it was
not updated during recording, triggering divergence.