Rework debugger init script handling to handle LLDB quirks.
The current hack of `script eval(...)` for LLDB requires our script to
use `lldb.debugger`. Some other unrelated hacks involving LLDB create
a situation where `lldb.debugger` is sometimes nondeterministically
set to `None` temporarily and asynchronously, during a short period
after debugger startup. The recommended way to run a script from a
file is `command script import`. Unfortunately `command script import`
requires the script file basename to be a valid Python identifer plus
the `.py` suffix, so we can't use our proc-pid scheme anymore.
Instead we now make the init script be responsible for deleting itself
when it finishes running.

We also have to reformat the LLDB init script to put the code
that needs to run eagerly in a `__lldb_init_module()` function. This
function takes a `debugger` argument so we can avoid using
`lldb.debugger`, which is the ultimate point of this change.