Don't leak extra fds into recorded processes
We are pretty good about setting CLOEXEC on most file descriptors.
However, we were missing them on the pagemap fd and the tty fd.
With lots of nested detaching, this can start filling up fd space,
which manifested as a CI failure on julia CI: https://github.com/JuliaLang/julia/issues/58979

Fix this by setting CLOEXEC as appropriate and add a test to make sure
this doesn't regress.