Support for C++>=20
Changes the CMake so that we can configure what c++ version we want to build with instead of hardcoding -std=c++17.

The part about turning off extensions, is because without it, rr is compiled with -std=gnu++17/20/23, instead of -std=c++17/... which would not match the previous hard coding.

If a user tries configuring with anything lower than c++17, it will fail at configuration time, instead of at build time.