log: Clear log_modules after changing log spec
I was expecting to be able to change the logging of a running rr
process by running `rr::apply_log_spec` at runtime under gdb, but
this did not quite work, because unlike `set_all_logging`, it does
not clear the log_modules cache. I think it would be useful for
this to work, so clear log_modules at the end of this function.